+++ core/src/ViewState/getBlockByPath.ts
@@ -1,0 +1,7 @@
+import { ViewState, Block } from '../types'
+
+export default function getBlockByPath (state: ViewState, path: number[]): Block {
+  return path.reduce((acc: any, val) => {
+    return acc.blocks[val]
+  }, state)
+}